+++ typescript/src/gapic-generator-typescript.ts
@@ -83,6 +83,10 @@ yargs.describe(
   'Set to true if GAPIC metadata generation is requested'
 );
 yargs.boolean('metadata');
+yargs.describe(
+  'transport',
+  'Default transport is gRPC. Set transport=rest for an API requires HTTP transport, or Google Discovery API.'
+);
 yargs.describe('protoc', 'Path to protoc binary');
 yargs.usage('Usage: $0 -I /path/to/googleapis');
 yargs.usage('  --output_dir /path/to/output_directory');
@@ -107,6 +111,7 @@ export interface IArguments {
   protoDirs?: string[];
   commonProtoPath?: string;
   descriptor?: string;
+  transport?: string;
   _: string[];
   $0: string;
 }
@@ -125,6 +130,7 @@ const template = argv.template as string | undefined;
 const gapicValidatorOut = argv.gapicValidatorOut as string | undefined;
 const validation = (argv.validation as string | undefined) ?? 'true';
 const metadata = argv.metadata as boolean | undefined;
+const transport = argv.transport as string | undefined;
 const protoc = (argv.protoc as string | undefined) ?? 'protoc';
 const protoDirs: string[] = [];
 if (argv.I) {
@@ -178,6 +184,9 @@ if (template) {
 }
 if (metadata) {
   protocCommand.push('--typescript_gapic_opt="metadata"');
+}
+if (transport && transport === 'rest') {
+  protocCommand.push('--typescript_gapic_opt="transport=rest"');
 }
 protocCommand.push(...protoDirsArg);
 protocCommand.push(...protoFiles);