+++ typescript/src/generator.ts
@@ -70,6 +70,7 @@ export class Generator {
   serviceYaml?: ServiceYaml;
   templates: string[];
   metadata?: boolean;
+  rest?: boolean;
 
   constructor() {
     this.request = protos.google.protobuf.compiler.CodeGeneratorRequest.create();
@@ -194,6 +195,12 @@ export class Generator {
     }
   }
 
+  private readRest() {
+    if (this.paramMap['transport'] === 'rest') {
+      this.rest = true;
+    }
+  }
+
   async initializeFromStdin() {
     const inputBuffer = await getStdin();
     this.request = protos.google.protobuf.compiler.CodeGeneratorRequest.decode(
@@ -207,6 +214,7 @@ export class Generator {
       this.readPublishPackageName();
       this.readMainServiceName();
       this.readTemplates();
+      this.readRest();
     }
   }
 
@@ -244,6 +252,7 @@ export class Generator {
       publishName: this.publishName,
       mainServiceName: this.mainServiceName,
       serviceYaml: this.serviceYaml,
+      rest: this.rest,
     });
     return api;
   }